refactor : controller, service interface remove
